Version 4 Release
Version 4 of the widget app now contains another page called the Translate page. User can put any text in any language in the input field of this page and can select a language to see the put text's translation on that selected language. This widget is very straight forward, but usefull in a lot of sense.